requirements.txt
aniso8601==7.0.0
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-GraphQL==2.0.1
Flask-SQLAlchemy==2.5.1
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.2.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.2
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pkg-resources==0.0.0
promise==2.3
PyMySQL==1.0.2
python-dotenv==0.20.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.36
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
